refactor: tidy stores
