+++ src/lib/stores/search-history.ts
